the hack number
Problem Name
The hack number

Problem Statement
You are given two strings S1 and S2, a character C (which can be either Y or N), and an integer I.
A Hack Number is defined as follows:
1. It is the first occurrence (the index of the first character) of the string S2 in the string S1 starting from the initial position I or later in S1, depending on the value of C.
2. If C is Y, then the matched S2 must have a space to the left of it in S1 or should   
be at the start of S1 It must have a space to the right of it in S1, or should be at
the end of S1.
3. If C is N, then there is no such compulsion and it can also be
inside a word of S1. (See Sample Test case 2 and 3 as an example)
4. If no element in S1 matches any element in S2, print “Goodbye Watson”.
Input format
First line: T (number of test cases)
For each test case
1. First line:S1
2. Second line:S2
3. Third line: C
4. Fourth line: I
Output format
Print the hack number for each test case. If no valid number exists, print “Goodbye Watson”.
Note
All indexes are 0-based, including 'I' and the answer value of 'Hack Number'.
Constraints
1 ≤ T ≤ 100
1 ≤ |S1| ≤ 50, containing letters('a'-'z', 'A'-'Z') and spaces.
1 ≤ |S2| ≤ 50, containing only letters('a'-'z', 'A'-'Z').
C will either be 'Y' or 'N'
I will be between 0 and L-1, L=length of S1.


















Sample Input
3
We love to hack on hackerearth
hack
Y
0
We love to hack on hackerearth
hack
Y
14
We love to hack on hackerearth
hack
N
14

Sample Output
11
Goodbye Watson.
19